Added wrapper element to featured images
Right now the get_the_image function doesn’t allow for a class to be
applied to the anchor tag surrounding the image. This make styling the
featured image a little annoying. Some of Justin’s themes add a class
to the anchor using JS, but this seems kind of ridiculous. I think I
prefer just adding a wrapper div for now.