Make accessible menu script an external jQuery plugin
Makes the JS work as expected, including the previously broken delay on mouseout.
(delay() only works with animations).

Also adds a check before loading fitvids to ensure it exists.

Changes copy:js grunt task to concat:js so that new plugin is preprended to theme.js file.
On this commit, the fitvids file is not copied from assets to theme, so it won't work. To be fixed
in the next commit by also concatenating that in too.

Skiplinks and mobile menu button to be moved to their own external plugins over time.

See #2