Enqueue Genericons separately
While I think it’s great  to save http requests wherever possible, in
this case I don’t think it’s worth it. There are a number of popular
plugins using Genericons which means we would be loading it twice if
any of them are active. The most noteworthy example of this is Jetpack.

Because of the likelihood that Automattic will change the directory
structure of the Genericons project, I’m also going to avoid pulling
the icons in via Bower for now. Related to #21