Reordered default widget areas
While it’s nice to have the header right widget area load first, it
also causes new installs to dump all the default widgets into it rather
than the primary sidebar. This is really annoying in Genesis and I’d
like to avoid it.

Also, because the header right area is likely to not have a title, It
is probably more semantically correct to use a div for widget wrappers
here.