Moved #site-inner and sidebar into main templates
Loading these from the header and footer templates cuts down on code
duplication, but makes the main templates less flexible. With these
elements in the header and footer, we’re limited in what can be placed
inside and outside of the main site-inner element.

This could also be handled by hooking things a bit more, but the goal
of Compass is to allow for maximum flexibility without forcing the use
of hooks. We only want to use the hooks sparingly in our themes and
more extensively in any supporting plugins we create.